Revert "Refactor tests before adding new JSON preprocessor (#22)"
This reverts commit 363680887b39f6bef58e40d6cbe6ffcad3145aae.
