+++ tests/site/newsletter/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture(scope="package")
+def dummy_email() -> str:
+    return "dummy@dummydomain.com"
